@@ -1,13 +1,13 @@
 /*
- * 3gpp-traffic-influence
+ * Nbsf_Management
  *
- * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+ * Binding Support Management Service API. © 2020, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
  *
- * API version: 1.0.1
+ * API version: 1.1.0
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package TrafficInfluence
+package Nbsf_Management
 
 import (
 	"context"
@@ -26,37 +26,163 @@ var (
 	_ context.Context
 )
 
-type TrafficInfluenceAPISubscriptionLevelPUTOperationApiService service
+type IndividualPCFBindingDocumentApiService service
 
 /*
-TrafficInfluenceAPISubscriptionLevelPUTOperationApiService Updates/replaces an existing subscription resource
+IndividualPCFBindingDocumentApiService Delete an existing Individual PCF Binding information
  * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param afId Identifier of the AF
- * @param subscriptionId Identifier of the subscription resource
- * @param trafficInfluSub Parameters to update/replace the existing subscription
-@return models.TrafficInfluSub
+ * @param bindingId Represents the individual PCF Session Binding.
 */
 
-func (a *TrafficInfluenceAPISubscriptionLevelPUTOperationApiService) AfIdSubscriptionsSubscriptionIdPut(ctx context.Context, afId string, subscriptionId string, trafficInfluSub models.TrafficInfluSub) (models.TrafficInfluSub, *http.Response, error) {
+func (a *IndividualPCFBindingDocumentApiService) DeleteIndPCFBinding(ctx context.Context, bindingId string) (*http.Response, error) {
 	var (
-		localVarHTTPMethod   = strings.ToUpper("Put")
+		localVarHTTPMethod   = strings.ToUpper("Delete")
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
 		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluSub
 	)
 
 	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/{afId}/subscriptions/{subscriptionId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"afId"+"}", fmt.Sprintf("%v", afId), -1)
-	localVarPath = strings.Replace(localVarPath, "{"+"subscriptionId"+"}", fmt.Sprintf("%v", subscriptionId), -1)
+	localVarPath := a.client.cfg.BasePath() + "/pcfBindings/{bindingId}"
+	localVarPath = strings.Replace(localVarPath, "{"+"bindingId"+"}", fmt.Sprintf("%v", bindingId), -1)
 
 	localVarHeaderParams := make(map[string]string)
 	localVarQueryParams := url.Values{}
 	localVarFormParams := url.Values{}
 
 	localVarHTTPContentTypes := []string{"application/json"}
+
+	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
+
+	// to determine the Accept header
+	localVarHTTPHeaderAccepts := []string{"application/problem+json"}
+
+	// set Accept header
+	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
+	if localVarHTTPHeaderAccept != "" {
+		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
+	}
+
+	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
+	if err != nil {
+		return nil, err
+	}
+
+	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
+	if err != nil || localVarHTTPResponse == nil {
+		return localVarHTTPResponse, err
+	}
+
+	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarHTTPResponse.Body.Close()
+	if err != nil {
+		return localVarHTTPResponse, err
+	}
+
+	apiError := openapi.GenericOpenAPIError{
+		RawBody:     localVarBody,
+		ErrorStatus: localVarHTTPResponse.Status,
+	}
+
+	switch localVarHTTPResponse.StatusCode {
+	case 204:
+		return localVarHTTPResponse, nil
+	case 400:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	case 401:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	case 403:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	case 404:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	case 429:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	case 500:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	case 503:
+		var v models.ProblemDetails
+		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
+		if err != nil {
+			apiError.ErrorStatus = err.Error()
+			return localVarHTTPResponse, apiError
+		}
+		apiError.ErrorModel = v
+		return localVarHTTPResponse, apiError
+	default:
+		return localVarHTTPResponse, nil
+	}
+}
+
+/*
+IndividualPCFBindingDocumentApiService Update an existing Individual PCF Binding information
+ * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ * @param bindingId Represents the individual PCF Session Binding.
+ * @param pcfBindingPatch Parameters to update the existing session binding
+@return PcfBinding
+*/
+
+func (a *IndividualPCFBindingDocumentApiService) UpdateIndPCFBinding(ctx context.Context, bindingId string, pcfBindingPatch models.PcfBindingPatch) (models.PcfBinding, *http.Response, error) {
+	var (
+		localVarHTTPMethod   = strings.ToUpper("Patch")
+		localVarPostBody     interface{}
+		localVarFormFileName string
+		localVarFileName     string
+		localVarFileBytes    []byte
+		localVarReturnValue  models.PcfBinding
+	)
+
+	// create path and map variables
+	localVarPath := a.client.cfg.BasePath() + "/pcfBindings/{bindingId}"
+	localVarPath = strings.Replace(localVarPath, "{"+"bindingId"+"}", fmt.Sprintf("%v", bindingId), -1)
+
+	localVarHeaderParams := make(map[string]string)
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
+
+	localVarHTTPContentTypes := []string{"application/merge-patch+json"}
 
 	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
 
@@ -70,7 +196,7 @@ func (a *TrafficInfluenceAPISubscriptionLevelPUTOperationApiService) AfIdSubscri
 	}
 
 	// body params
-	localVarPostBody = &trafficInfluSub
+	localVarPostBody = &pcfBindingPatch
 
 	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
 	if err != nil {
