@@ -1,26 +1,23 @@
-//go:build !debug
-// +build !debug
-
 /*
- * Namf_MT
+ * Nbsf_Management
  *
- * AMF Mobile Termination Service
+ * Binding Support Management Service API. © 2020, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
  *
- * API version: 1.0.0
+ * API version: 1.1.0
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package Namf_MT
+package Nbsf_Management
 
-// APIClient manages communication with the Namf_MT API v1.0.0
+// APIClient manages communication with the Nbsf_Management API v1.1.0
 // In most cases there should be only one, shared, APIClient.
 type APIClient struct {
 	cfg    *Configuration
 	common service // Reuse a single struct instead of allocating one for each service on the heap.
 
 	// API Services
-	UeContextDocumentApi  *UeContextDocumentApiService
-	UeReachIndDocumentApi *UeReachIndDocumentApiService
+	IndividualPCFBindingDocumentApi *IndividualPCFBindingDocumentApiService
+	PCFBindingsCollectionApi        *PCFBindingsCollectionApiService
 }
 
 type service struct {
@@ -35,8 +32,8 @@ func NewAPIClient(cfg *Configuration) *APIClient {
 	c.common.client = c
 
 	// API Services
-	c.UeContextDocumentApi = (*UeContextDocumentApiService)(&c.common)
-	c.UeReachIndDocumentApi = (*UeReachIndDocumentApiService)(&c.common)
+	c.IndividualPCFBindingDocumentApi = (*IndividualPCFBindingDocumentApiService)(&c.common)
+	c.PCFBindingsCollectionApi = (*PCFBindingsCollectionApiService)(&c.common)
 
 	return c
 }
