@@ -1,19 +1,24 @@
 /*
  * 3gpp-traffic-influence
  *
- * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+ * API for AF traffic influence © 2020, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved.
  *
- * API version: 1.0.1
+ * API version: 1.1.0
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package TrafficInfluence
+package Nnef_TrafficInfluence
 
 import (
 	"net/http"
 	"strings"
 )
 
+// contextKeys are used to identify the type of value in the context.
+// Since these are string, it is possible to get a short description of the
+// context key for logging and debugging using key.String().
+
+// Configuration stores the configuration of the API client
 type Configuration struct {
 	url           string
 	basePath      string
@@ -23,6 +28,7 @@ type Configuration struct {
 	httpClient    *http.Client
 }
 
+// NewConfiguration returns a new Configuration object
 func NewConfiguration() *Configuration {
 	cfg := &Configuration{
 		basePath:      "https://example.com/3gpp-traffic-influence/v1",
