@@ -33,1427 +33,6 @@ type DefaultApiService service
 /*
 DefaultApiService
  * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param optional nil or *ApplicationDataInfluenceDataGetParamOpts - Optional Parameters:
- * @param "InfluenceIds" (optional.Interface of []string) -  Each element identifies a service.
- * @param "Dnns" (optional.Interface of []string) -  Each element identifies a DNN.
- * @param "Snssais" (optional.Interface of []models.Snssai) -  Each element identifies a slice.
- * @param "InternalGroupIds" (optional.Interface of []string) -  Each element identifies a group of users.
- * @param "Supis" (optional.Interface of []string) -  Each element identifies the user.
-@return []models.TrafficInfluData
-*/
-
-type ApplicationDataInfluenceDataGetParamOpts struct {
-	InfluenceIds     optional.Interface
-	Dnns             optional.Interface
-	Snssais          optional.Interface
-	InternalGroupIds optional.Interface
-	Supis            optional.Interface
-}
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataGet(ctx context.Context, localVarOptionals *ApplicationDataInfluenceDataGetParamOpts) ([]models.TrafficInfluData, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Get")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  []models.TrafficInfluData
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData"
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	if localVarOptionals != nil && localVarOptionals.InfluenceIds.IsSet() {
-		localVarQueryParams.Add("influence-Ids", openapi.ParameterToString(localVarOptionals.InfluenceIds.Value(), "csv"))
-	}
-	if localVarOptionals != nil && localVarOptionals.Dnns.IsSet() {
-		localVarQueryParams.Add("dnns", openapi.ParameterToString(localVarOptionals.Dnns.Value(), "csv"))
-	}
-	if localVarOptionals != nil && localVarOptionals.Snssais.IsSet() {
-		localVarQueryParams.Add("snssais", openapi.ParameterToString(localVarOptionals.Snssais.Value(), "csv"))
-	}
-	if localVarOptionals != nil && localVarOptionals.InternalGroupIds.IsSet() {
-		localVarQueryParams.Add("internal-Group-Ids", openapi.ParameterToString(localVarOptionals.InternalGroupIds.Value(), "csv"))
-	}
-	if localVarOptionals != nil && localVarOptionals.Supis.IsSet() {
-		localVarQueryParams.Add("supis", openapi.ParameterToString(localVarOptionals.Supis.Value(), "csv"))
-	}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 200:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 406:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 414:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param influenceId The Identifier of an Individual Influence Data to be updated.It shall apply the format of Data type string.
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataInfluenceIdDelete(ctx context.Context, influenceId string) (*http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Delete")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/{influenceId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"influenceId"+"}", fmt.Sprintf("%v", influenceId), -1)
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 204:
-		return localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	default:
-		return localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param influenceId The Identifier of an Individual Influence Data to be updated.It shall apply the format of Data type string.
- * @param trafficInfluDataPatch
-@return models.TrafficInfluData
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataInfluenceIdPatch(ctx context.Context, influenceId string, trafficInfluDataPatch models.TrafficInfluDataPatch) (models.TrafficInfluData, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Patch")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluData
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/{influenceId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"influenceId"+"}", fmt.Sprintf("%v", influenceId), -1)
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	// body params
-	localVarPostBody = &trafficInfluDataPatch
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 200:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 204:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 411:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 413:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 415:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param influenceId The Identifier of an Individual Influence Data to be created or updated.It shall apply the format of Data type string.
- * @param trafficInfluData
-@return models.TrafficInfluData
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataInfluenceIdPut(ctx context.Context, influenceId string, trafficInfluData models.TrafficInfluData) (models.TrafficInfluData, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Put")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluData
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/{influenceId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"influenceId"+"}", fmt.Sprintf("%v", influenceId), -1)
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	// body params
-	localVarPostBody = &trafficInfluData
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 201:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 200:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 204:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 411:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 413:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 414:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 415:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param optional nil or *ApplicationDataInfluenceDataSubsToNotifyGetParamOpts - Optional Parameters:
- * @param "Dnn" (optional.String) -  Identifies a DNN.
- * @param "Snssai" (optional.Interface of models.Snssai) -  Identifies a slice.
- * @param "InternalGroupId" (optional.String) -  Identifies a group of users.
- * @param "Supi" (optional.String) -  Identifies a user.
-@return []models.TrafficInfluSub
-*/
-
-type ApplicationDataInfluenceDataSubsToNotifyGetParamOpts struct {
-	Dnn             optional.String
-	Snssai          optional.Interface
-	InternalGroupId optional.String
-	Supi            optional.String
-}
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataSubsToNotifyGet(ctx context.Context, localVarOptionals *ApplicationDataInfluenceDataSubsToNotifyGetParamOpts) ([]models.TrafficInfluSub, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Get")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  []models.TrafficInfluSub
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/subs-to-notify"
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	if localVarOptionals != nil && localVarOptionals.Dnn.IsSet() && localVarOptionals.Dnn.Value() != "" {
-		localVarQueryParams.Add("dnn", openapi.ParameterToString(localVarOptionals.Dnn.Value(), ""))
-	}
-	if localVarOptionals != nil && localVarOptionals.Snssai.IsSet() {
-		localVarQueryParams.Add("snssai", openapi.ParameterToString(localVarOptionals.Snssai.Value(), ""))
-	}
-	if localVarOptionals != nil && localVarOptionals.InternalGroupId.IsSet() && localVarOptionals.InternalGroupId.Value() != "" {
-		localVarQueryParams.Add("internal-Group-Id", openapi.ParameterToString(localVarOptionals.InternalGroupId.Value(), ""))
-	}
-	if localVarOptionals != nil && localVarOptionals.Supi.IsSet() && localVarOptionals.Supi.Value() != "" {
-		localVarQueryParams.Add("supi", openapi.ParameterToString(localVarOptionals.Supi.Value(), ""))
-	}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 200:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 406:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 414:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param trafficInfluSub
-@return models.TrafficInfluSub
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataSubsToNotifyPost(ctx context.Context, trafficInfluSub models.TrafficInfluSub) (models.TrafficInfluSub, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Post")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluSub
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/subs-to-notify"
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	// body params
-	localVarPostBody = &trafficInfluSub
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 201:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 411:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 413:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 415:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param subscriptionId String identifying a subscription to the Individual Influence Data Subscription
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataSubsToNotifySubscriptionIdDelete(ctx context.Context, subscriptionId string) (*http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Delete")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/subs-to-notify/{subscriptionId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"subscriptionId"+"}", fmt.Sprintf("%v", subscriptionId), -1)
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 204:
-		return localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarHTTPResponse, apiError
-	default:
-		return localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param subscriptionId String identifying a subscription to the Individual Influence Data Subscription
-@return models.TrafficInfluSub
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataSubsToNotifySubscriptionIdGet(ctx context.Context, subscriptionId string) (models.TrafficInfluSub, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Get")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluSub
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/subs-to-notify/{subscriptionId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"subscriptionId"+"}", fmt.Sprintf("%v", subscriptionId), -1)
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 200:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 406:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 414:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param subscriptionId String identifying a subscription to the Individual Influence Data Subscription
- * @param trafficInfluSub
-@return models.TrafficInfluSub
-*/
-
-func (a *DefaultApiService) ApplicationDataInfluenceDataSubsToNotifySubscriptionIdPut(ctx context.Context, subscriptionId string, trafficInfluSub models.TrafficInfluSub) (models.TrafficInfluSub, *http.Response, error) {
-	var (
-		localVarHTTPMethod   = strings.ToUpper("Put")
-		localVarPostBody     interface{}
-		localVarFormFileName string
-		localVarFileName     string
-		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluSub
-	)
-
-	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData/subs-to-notify/{subscriptionId}"
-	localVarPath = strings.Replace(localVarPath, "{"+"subscriptionId"+"}", fmt.Sprintf("%v", subscriptionId), -1)
-
-	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := url.Values{}
-	localVarFormParams := url.Values{}
-
-	localVarHTTPContentTypes := []string{"application/json"}
-
-	localVarHeaderParams["Content-Type"] = localVarHTTPContentTypes[0] // use the first content type specified in 'consumes'
-
-	// to determine the Accept header
-	localVarHTTPHeaderAccepts := []string{"application/json", "application/problem+json"}
-
-	// set Accept header
-	localVarHTTPHeaderAccept := openapi.SelectHeaderAccept(localVarHTTPHeaderAccepts)
-	if localVarHTTPHeaderAccept != "" {
-		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
-	}
-
-	// body params
-	localVarPostBody = &trafficInfluSub
-
-	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
-	if err != nil {
-		return localVarReturnValue, nil, err
-	}
-
-	localVarHTTPResponse, err := openapi.CallAPI(a.client.cfg, r)
-	if err != nil || localVarHTTPResponse == nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	localVarBody, err := ioutil.ReadAll(localVarHTTPResponse.Body)
-	localVarHTTPResponse.Body.Close()
-	if err != nil {
-		return localVarReturnValue, localVarHTTPResponse, err
-	}
-
-	apiError := openapi.GenericOpenAPIError{
-		RawBody:     localVarBody,
-		ErrorStatus: localVarHTTPResponse.Status,
-	}
-
-	switch localVarHTTPResponse.StatusCode {
-	case 200:
-		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-		}
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 204:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	case 400:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 401:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 403:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 404:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 411:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 413:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 415:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 429:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 500:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 503:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	default:
-		return localVarReturnValue, localVarHTTPResponse, nil
-	}
-}
-
-/*
-DefaultApiService
- * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
  * @param appId Indicate the application identifier for the request pfd(s).It shall apply the format of Data type ApplicationId.
 */
 
@@ -4013,7 +2592,7 @@ func (a *DefaultApiService) PolicyDataUesUeIdOperatorSpecificDataGet(ctx context
 	localVarQueryParams := url.Values{}
 	localVarFormParams := url.Values{}
 
-	if localVarOptionals != nil && localVarOptionals.Fields.IsSet() {
+	if localVarOptionals != nil && localVarOptionals.Fields.IsSet() && localVarOptionals.Fields.Value() != "" {
 		localVarQueryParams.Add("fields", openapi.ParameterToString(localVarOptionals.Fields.Value(), "csv"))
 	}
 
@@ -4494,13 +3073,13 @@ func (a *DefaultApiService) PolicyDataUesUeIdSmDataGet(ctx context.Context, ueId
 	localVarQueryParams := url.Values{}
 	localVarFormParams := url.Values{}
 
-	if localVarOptionals != nil && localVarOptionals.Snssai.IsSet() {
+	if localVarOptionals != nil && localVarOptionals.Snssai.IsSet() && localVarOptionals.Snssai.Value() != "" {
 		localVarQueryParams.Add("snssai", openapi.ParameterToString(localVarOptionals.Snssai.Value(), ""))
 	}
-	if localVarOptionals != nil && localVarOptionals.Dnn.IsSet() {
+	if localVarOptionals != nil && localVarOptionals.Dnn.IsSet() && localVarOptionals.Dnn.Value() != "" {
 		localVarQueryParams.Add("dnn", openapi.ParameterToString(localVarOptionals.Dnn.Value(), ""))
 	}
-	if localVarOptionals != nil && localVarOptionals.Fields.IsSet() {
+	if localVarOptionals != nil && localVarOptionals.Fields.IsSet() && localVarOptionals.Fields.Value() != "" {
 		localVarQueryParams.Add("fields", openapi.ParameterToString(localVarOptionals.Fields.Value(), "csv"))
 	}
 
