@@ -1,24 +1,25 @@
 /*
- * 3gpp-traffic-influence
+ * Nudr_DataRepository API OpenAPI file
  *
- * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+ * Unified Data Repository Service
  *
- * API version: 1.0.1
+ * API version: 1.0.0
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package TrafficInfluence
+package Nudr_DataRepository
 
 import (
+	"github.com/free5gc/openapi"
+	"github.com/free5gc/openapi/models"
+
 	"context"
-	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strings"
 
-	"github.com/free5gc/openapi"
-	"github.com/free5gc/openapi/models"
+	"github.com/antihax/optional"
 )
 
 // Linger please
@@ -26,33 +27,60 @@ var (
 	_ context.Context
 )
 
-type TrafficInfluenceAPISubscriptionLevelPOSTOperationApiService service
+type InfluenceDataApiService service
 
 /*
-TrafficInfluenceAPISubscriptionLevelPOSTOperationApiService Creates a new subscription resource
+InfluenceDataApiService
  * @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param afId Identifier of the AF
- * @param trafficInfluSub Request to create a new subscription resource
-@return models.TrafficInfluSub
+ * @param optional nil or *ApplicationDataInfluenceDataGetParamOpts - Optional Parameters:
+ * @param "InfluenceIds" (optional.Interface of []string) -  Each element identifies a service.
+ * @param "Dnns" (optional.Interface of []string) -  Each element identifies a DNN.
+ * @param "Snssais" (optional.Interface of []models.Snssai) -  Each element identifies a slice.
+ * @param "InternalGroupIds" (optional.Interface of []string) -  Each element identifies a group of users.
+ * @param "Supis" (optional.Interface of []string) -  Each element identifies the user.
+@return []models.TrafficInfluData
 */
 
-func (a *TrafficInfluenceAPISubscriptionLevelPOSTOperationApiService) AfIdSubscriptionsPost(ctx context.Context, afId string, trafficInfluSub models.TrafficInfluSub) (models.TrafficInfluSub, *http.Response, error) {
+type ApplicationDataInfluenceDataGetParamOpts struct {
+	InfluenceIds     optional.Interface
+	Dnns             optional.Interface
+	Snssais          optional.Interface
+	InternalGroupIds optional.Interface
+	Supis            optional.Interface
+}
+
+func (a *InfluenceDataApiService) ApplicationDataInfluenceDataGet(ctx context.Context, localVarOptionals *ApplicationDataInfluenceDataGetParamOpts) ([]models.TrafficInfluData, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = strings.ToUpper("Post")
+		localVarHTTPMethod   = strings.ToUpper("Get")
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
 		localVarFileBytes    []byte
-		localVarReturnValue  models.TrafficInfluSub
+		localVarReturnValue  []models.TrafficInfluData
 	)
 
 	// create path and map variables
-	localVarPath := a.client.cfg.BasePath() + "/{afId}/subscriptions"
-	localVarPath = strings.Replace(localVarPath, "{"+"afId"+"}", fmt.Sprintf("%v", afId), -1)
+	localVarPath := a.client.cfg.BasePath() + "/application-data/influenceData"
 
 	localVarHeaderParams := make(map[string]string)
 	localVarQueryParams := url.Values{}
 	localVarFormParams := url.Values{}
+
+	if localVarOptionals != nil && localVarOptionals.InfluenceIds.IsSet() && localVarOptionals.InfluenceIds.Value() != "" {
+		localVarQueryParams.Add("influence-Ids", openapi.ParameterToString(localVarOptionals.InfluenceIds.Value(), "csv"))
+	}
+	if localVarOptionals != nil && localVarOptionals.Dnns.IsSet() && localVarOptionals.Dnns.Value() != "" {
+		localVarQueryParams.Add("dnns", openapi.ParameterToString(localVarOptionals.Dnns.Value(), "csv"))
+	}
+	if localVarOptionals != nil && localVarOptionals.Snssais.IsSet() && localVarOptionals.Snssais.Value() != "" {
+		localVarQueryParams.Add("snssais", openapi.ParameterToString(localVarOptionals.Snssais.Value(), "csv"))
+	}
+	if localVarOptionals != nil && localVarOptionals.InternalGroupIds.IsSet() && localVarOptionals.InternalGroupIds.Value() != "" {
+		localVarQueryParams.Add("internal-Group-Ids", openapi.ParameterToString(localVarOptionals.InternalGroupIds.Value(), "csv"))
+	}
+	if localVarOptionals != nil && localVarOptionals.Supis.IsSet() && localVarOptionals.Supis.Value() != "" {
+		localVarQueryParams.Add("supis", openapi.ParameterToString(localVarOptionals.Supis.Value(), "csv"))
+	}
 
 	localVarHTTPContentTypes := []string{"application/json"}
 
@@ -66,9 +94,6 @@ func (a *TrafficInfluenceAPISubscriptionLevelPOSTOperationApiService) AfIdSubscr
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
-
-	// body params
-	localVarPostBody = &trafficInfluSub
 
 	r, err := openapi.PrepareRequest(ctx, a.client.cfg, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, localVarFormFileName, localVarFileName, localVarFileBytes)
 	if err != nil {
@@ -92,7 +117,7 @@ func (a *TrafficInfluenceAPISubscriptionLevelPOSTOperationApiService) AfIdSubscr
 	}
 
 	switch localVarHTTPResponse.StatusCode {
-	case 201:
+	case 200:
 		err = openapi.Deserialize(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 		if err != nil {
 			apiError.ErrorStatus = err.Error()
@@ -134,25 +159,9 @@ func (a *TrafficInfluenceAPISubscriptionLevelPOSTOperationApiService) AfIdSubscr
 		}
 		apiError.ErrorModel = v
 		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 411:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 413:
-		var v models.ProblemDetails
-		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
-		if err != nil {
-			apiError.ErrorStatus = err.Error()
-			return localVarReturnValue, localVarHTTPResponse, apiError
-		}
-		apiError.ErrorModel = v
-		return localVarReturnValue, localVarHTTPResponse, apiError
-	case 415:
+	case 406:
+		return localVarReturnValue, localVarHTTPResponse, nil
+	case 414:
 		var v models.ProblemDetails
 		err = openapi.Deserialize(&v, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 		if err != nil {
