@@ -1,7 +1,7 @@
 /*
- * NRF NFManagement Service
+ * Nbsf_Management
  *
- * NRF NFManagement Service
+ * Binding Support Management Service API
  *
  * API version: 1.0.1
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
