@@ -1,7 +1,7 @@
 /*
- * 3gpp-traffic-influence
+ * Nbsf_Management
  *
- * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+ * Binding Support Management Service API
  *
  * API version: 1.0.1
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
