@@ -21,19 +21,40 @@ type TrafficInfluData struct {
 	// Identifies an application.
 	AfAppId string `json:"afAppId,omitempty" bson:"afAppId"`
 	Dnn     string `json:"dnn,omitempty" bson:"dnn"`
-	// Identifies Ethernet packet filters.Either \"trafficFilters\" or \"ethTrafficFilters\" shall be included if applicable.
+	// Identifies Ethernet packet filters. Either \"trafficFilters\" or \"ethTrafficFilters\" shall be included if applicable.
 	EthTrafficFilters []EthFlowDescription `json:"ethTrafficFilters,omitempty" bson:"ethTrafficFilters"`
-	Snssai            *Snssai              `json:"snssai,omitempty" bson:"snssai"`
+	Snssai            *Snssai `json:"snssai,omitempty" bson:"snssai"`
 	// Identifies a group of users.
 	InterGroupId string `json:"interGroupId,omitempty" bson:"interGroupId"`
 	Supi         string `json:"supi,omitempty" bson:"supi"`
-	// Identifies IP packet filters.Either \"trafficFilters\" or \"ethTrafficFilters\" shall be included if applicable.
+	// Identifies IP packet filters. Either \"trafficFilters\" or \"ethTrafficFilters\" shall be included if applicable.
 	TrafficFilters []FlowInfo `json:"trafficFilters,omitempty" bson:"trafficFilters"`
 	// Identifies the N6 traffic routing requirement.
-	TrafficRoutes  []RouteToLocation `json:"trafficRoutes" bson:"trafficRoutes"`
-	ValidStartTime *time.Time        `json:"validStartTime,omitempty" bson:"validStartTime"`
-	ValidEndTime   *time.Time        `json:"validEndTime,omitempty" bson:"validEndTime"`
-	NwAreaInfo     *NetworkAreaInfo  `json:"nwAreaInfo,omitempty" bson:"nwAreaInfo"`
+	TrafficRoutes []RouteToLocation `json:"trafficRoutes,omitempty" bson:"trafficRoutes"`
+
+	TraffCorreInd bool `json:"traffCorreInd,omitempty" bson:"traffCorreInd"`
+
+	ValidStartTime *time.Time `json:"validStartTime,omitempty" bson:"validStartTime"`
+
+	ValidEndTime *time.Time `json:"validEndTime,omitempty" bson:"validEndTime"`
+
+	// Identifies the temporal validities for the N6 traffic routing requirement.
+	TempValidities []TemporalValidity `json:"tempValidities,omitempty" bson:"tempValidities"`
+
+	NwAreaInfo *NetworkAreaInfo `json:"nwAreaInfo,omitempty" bson:"nwAreaInfo"`
+
 	// string providing an URI formatted according to IETF RFC 3986.
 	UpPathChgNotifUri string `json:"upPathChgNotifUri,omitempty" bson:"upPathChgNotifUri"`
+
+	SubscribedEvents []SubscribedEvent `json:"subscribedEvents,omitempty" bson:"subscribedEvents"`
+
+	DnaiChgType DnaiChangeType `json:"dnaiChgType,omitempty" bson:"dnaiChgType"`
+
+	AfAckInd bool `json:"afAckInd,omitempty" bson:"afAckInd"`
+
+	AddrPreserInd bool `json:"addrPreserInd,omitempty" bson:"addrPreserInd"`
+
+	SupportedFeatures string `json:"supportedFeatures,omitempty" bson:"supportedFeatures"`
+
+	ResUri string `json:"resUri,omitempty" bson:"resUri"`
 }
